UT: Add tests for new RegistrationStatus resource
